Moved utility value objects to own package
