+++ src/de/hawai/bicycle_tracking/server/astcore/customermanagement/IUser.java
@@ -2,8 +2,8 @@ package de.hawai.bicycle_tracking.server.astcore.customermanagement;
 
 import java.util.Date;
 
-import de.hawai.bicycle_tracking.server.utility.Address;
-import de.hawai.bicycle_tracking.server.utility.EMail;
+import de.hawai.bicycle_tracking.server.utility.value.Address;
+import de.hawai.bicycle_tracking.server.utility.value.EMail;
 
 public interface IUser {
